+++ src/apps/uwu-lend/contracts/ethers/factories/UwuLendStakingV1__factory.ts
@@ -4,7 +4,7 @@
 
 import { Contract, Signer, utils } from 'ethers';
 import type { Provider } from '@ethersproject/providers';
-import type { UwuLendMultiFeeV1, UwuLendMultiFeeV1Interface } from '../UwuLendMultiFeeV1';
+import type { UwuLendStakingV1, UwuLendStakingV1Interface } from '../UwuLendStakingV1';
 
 const _abi = [
   {
@@ -862,12 +862,12 @@ const _abi = [
   },
 ];
 
-export class UwuLendMultiFeeV1__factory {
+export class UwuLendStakingV1__factory {
   static readonly abi = _abi;
-  static createInterface(): UwuLendMultiFeeV1Interface {
-    return new utils.Interface(_abi) as UwuLendMultiFeeV1Interface;
+  static createInterface(): UwuLendStakingV1Interface {
+    return new utils.Interface(_abi) as UwuLendStakingV1Interface;
   }
-  static connect(address: string, signerOrProvider: Signer | Provider): UwuLendMultiFeeV1 {
-    return new Contract(address, _abi, signerOrProvider) as UwuLendMultiFeeV1;
+  static connect(address: string, signerOrProvider: Signer | Provider): UwuLendStakingV1 {
+    return new Contract(address, _abi, signerOrProvider) as UwuLendStakingV1;
   }
 }
