+++ src/apps/uwu-lend/contracts/index.ts
@@ -6,8 +6,8 @@ import { Network } from '~types/network.interface';
 
 import {
   UwuLendDataProvider__factory,
-  UwuLendMultiFeeV1__factory,
-  UwuLendMultiFeeV2__factory,
+  UwuLendStakingV1__factory,
+  UwuLendStakingV2__factory,
   UwuLendUToken__factory,
 } from './ethers';
 
@@ -23,11 +23,11 @@ export class UwuLendContractFactory extends ContractFactory {
   uwuLendDataProvider({ address, network }: ContractOpts) {
     return UwuLendDataProvider__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
-  uwuLendMultiFeeV1({ address, network }: ContractOpts) {
-    return UwuLendMultiFeeV1__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  uwuLendStakingV1({ address, network }: ContractOpts) {
+    return UwuLendStakingV1__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
-  uwuLendMultiFeeV2({ address, network }: ContractOpts) {
-    return UwuLendMultiFeeV2__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  uwuLendStakingV2({ address, network }: ContractOpts) {
+    return UwuLendStakingV2__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
   uwuLendUToken({ address, network }: ContractOpts) {
     return UwuLendUToken__factory.connect(address, this.appToolkit.getNetworkProvider(network));
@@ -35,6 +35,6 @@ export class UwuLendContractFactory extends ContractFactory {
 }
 
 export type { UwuLendDataProvider } from './ethers';
-export type { UwuLendMultiFeeV1 } from './ethers';
-export type { UwuLendMultiFeeV2 } from './ethers';
+export type { UwuLendStakingV1 } from './ethers';
+export type { UwuLendStakingV2 } from './ethers';
 export type { UwuLendUToken } from './ethers';
