+++ src/apps/uwu-lend/uwu-lend.module.ts
@@ -3,12 +3,10 @@ import { Module } from '@nestjs/common';
 import { AbstractApp } from '~app/app.dynamic-module';
 
 import { UwuLendContractFactory } from './contracts';
-import { EthereumUwuLendLpStakingV1ContractPositionFetcher } from './ethereum/uwu-lend.lp-staking-v1.contract-position-fetcher';
-import { EthereumUwuLendLpStakingV2ContractPositionFetcher } from './ethereum/uwu-lend.lp-staking-v2.contract-position-fetcher';
+import { EthereumUwuLendPlatformFeesV1PositionFetcher } from './ethereum/uwu-lend.platform-fees-v1.contract-position-fetcher';
+import { EthereumUwuLendPlatformFeesV2PositionFetcher } from './ethereum/uwu-lend.platform-fees-v2.contract-position-fetcher';
 import { EthereumUwuLendSupplyTokenFetcher } from './ethereum/uwu-lend.supply.token-fetcher';
 import { EthereumUwuLendVariableDebtTokenFetcher } from './ethereum/uwu-lend.variable-debt.token-fetcher';
-import { EthereumUwuLendVestingV1ContractPositionFetcher } from './ethereum/uwu-lend.vesting-v1.contract-position-fetcher';
-import { EthereumUwuLendVestingV2ContractPositionFetcher } from './ethereum/uwu-lend.vesting-v2.contract-position-fetcher';
 import { UwuLendAppDefinition } from './uwu-lend.definition';
 
 @Module({
@@ -17,10 +15,8 @@ import { UwuLendAppDefinition } from './uwu-lend.definition';
     UwuLendContractFactory,
     EthereumUwuLendSupplyTokenFetcher,
     EthereumUwuLendVariableDebtTokenFetcher,
-    EthereumUwuLendLpStakingV1ContractPositionFetcher,
-    EthereumUwuLendLpStakingV2ContractPositionFetcher,
-    EthereumUwuLendVestingV1ContractPositionFetcher,
-    EthereumUwuLendVestingV2ContractPositionFetcher,
+    EthereumUwuLendPlatformFeesV1PositionFetcher,
+    EthereumUwuLendPlatformFeesV2PositionFetcher,
   ],
 })
 export class UwuLendAppModule extends AbstractApp() {}
